pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solarathon"
description = "Template project for Solarathon 2023"
version = "0.0.1"
dependencies = [
    "solara==1.23.0",
"farm-haystack[faiss]",
"farm-haystack[inference]", 
"farm-haystack[preprocessing]",
"python-dotenv"
]

[project.optional-dependencies]
dev = [
    "mypy",
]

[tool.hatch.build]
include = [
  "**/*.css",
  "**/*.py",
  "**/*.json",
  "**/*.faiss",
  "**/*.db",
]
